Implement modular exponentiation using square-and-sum strat
This strikes a nice happy medium between the two extremes:
exponentiation followed by modulo (fast but very likely to overflow) and
modulation between every multiplication in the exponentiation (slow but
very unlikely to overflow); thus the max value is √T::MAX, not the `e`th root of
T::MAX (though not as good as T::MAX/2), and the the order notation is
O(log(e)), not O(e) (though not as good as O(1)).
TL;DR: fast-ish and safe-ish.
(NOTE: in Debug mode at least, fn will panic before returning wrong
result)